backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-dotenv>=1.0.1
pydantic>=2.7.4
aiofiles>=23.2.1
